convert array to map storage and added logrus to log errors

+++ command-line-local-shortener/main.go
@@ -5,50 +5,40 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
-	"log"
 	"net/http"
 	"net/url"
 	"os"
 
 	"github.com/gin-gonic/gin"
 	gonanoid "github.com/matoous/go-nanoid/v2"
+	"github.com/sirupsen/logrus"
 )
+
+var dataArray map[string]string
 
 func genID(length int) string {
 
 	id, err := gonanoid.New(length)
 	if err != nil {
-		panic(err)
+		logrus.Errorf("error while creating new id: %v", err)
+		os.Exit(1)
 	}
-	fmt.Printf("Generated id: %s\n", id)
 	return id
 }
 
 func isUrl(str string) bool {
 	u, err := url.Parse(str)
-	return err == nil && u.Scheme != "" && u.Host != ""
-}
-
-type urldatabase struct {
-	UrlID   string `json:"urlID"`
-	LongURL string `json:"longURL"`
+	if err != nil {
+		logrus.Errorf("error parsing url: %v", err)
+		os.Exit(1)
+	}
+	return u.Scheme != "" && u.Host != ""
 }
 
 func dataProccess() {
 	sc := bufio.NewScanner(os.Stdin)
 	idLength := 3
 	for {
-		var dataArray []urldatabase
-
-		jsonFile, err := os.Open("urlmap.json")
-		if err != nil {
-			fmt.Println(err)
-		}
-		fmt.Println("Successfully Opened url.json")
-		byteValue, _ := ioutil.ReadAll(jsonFile)
-		if err := json.Unmarshal(byteValue, &dataArray); err != nil {
-			log.Println(err)
-		}
 
 		sc.Scan()
 		input := sc.Text()
@@ -61,24 +51,23 @@ func dataProccess() {
 			if isUrl(inputURL) {
 				for {
 					currID := genID(idLength)
-					// if true {
-					dataArray = append(dataArray, urldatabase{UrlID: currID, LongURL: inputURL})
+					if _, exists := dataArray[currID]; !exists {
+						dataArray[currID] = inputURL
+						result, err := json.Marshal(dataArray)
+						if err != nil {
+							logrus.Errorf("error while marshalling json: %v", err)
+							os.Exit(1)
+						}
+						ioutil.WriteFile("urlmap.json", result, 0644)
 
-					result, err := json.Marshal(dataArray)
-					if err != nil {
-						log.Println(err)
+						for _, v := range dataArray {
+							fmt.Println(v)
+						}
+
+						break
+					} else {
+						idLength += 1
 					}
-					ioutil.WriteFile("urlmap.json", result, 0644)
-					jsonFile.Close()
-
-					for _, v := range dataArray {
-						fmt.Println(v)
-					}
-
-					break
-					// } else {
-					// 	idLength += 1
-					// }
 				}
 			} else {
 				fmt.Println("invalid url")
@@ -92,30 +81,30 @@ func dataProccess() {
 func getReq(c *gin.Context) {
 	id := c.Param("id")
 
-	var dataArray []urldatabase
+	fmt.Println(id)
+	url, exists := dataArray[id]
+	if exists {
+		c.Redirect(http.StatusMovedPermanently, url)
+	}
+}
+
+func main() {
 
 	jsonFile, err := os.Open("urlmap.json")
 	if err != nil {
-		fmt.Println(err)
+		logrus.Errorf("error while opening map file: %v", err)
+		os.Exit(1)
 	}
+	defer jsonFile.Close()
+	fmt.Println("Successfully Opened url.json")
 	byteValue, _ := ioutil.ReadAll(jsonFile)
 	if err := json.Unmarshal(byteValue, &dataArray); err != nil {
-		log.Println(err)
+		logrus.Errorf("error while unmarshalling json: %v", err)
+		os.Exit(1)
 	}
 
-	jsonFile.Close()
-	fmt.Println(id)
-	for i := range dataArray {
-		if dataArray[i].UrlID == id {
-			fmt.Println("yes")
-			c.Redirect(http.StatusMovedPermanently, dataArray[i].LongURL)
-			break
-		}
-	}
+	dataProccess()
 
-}
-
-func routerSetup() *gin.Engine {
 	router := gin.Default()
 
 	router.GET("/:id", getReq)
@@ -124,13 +113,6 @@ func routerSetup() *gin.Engine {
 		c.JSON(200, gin.H{"hello": "world"})
 	})
 
-	return router
-}
-
-func main() {
-	dataProccess()
-
-	router := routerSetup()
-
 	router.Run(":8090")
+
 }
